requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
chromadb==0.5.5
sentence-transformers==3.2.0
docling==2.5.0
pydantic==2.9.2
google-generativeai==0.8.3
